added automatic image deletion

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const mongoose = require('mongoose')
 const { validationResult } = require('express-validator')
 
+const fileHelper = require('../util/file')
 const Product = require('../models/product')
 
 exports.getAddProduct = (req, res, next) => {
@@ -113,7 +114,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId
   const updatedTitle = req.body.title
   const updatedPrice = req.body.price
-  const image  = req.file
+  const image = req.file
   const updatedDesc = req.body.description
 
   const errors = validationResult(req)
@@ -135,7 +136,6 @@ exports.postEditProduct = (req, res, next) => {
     })
   }
 
-
   Product.findById(prodId)
     .then((product) => {
       if (product.userId.toString() !== req.user._id.toString()) {
@@ -144,6 +144,7 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle
       product.price = updatedPrice
       if (image) {
+        fileHelper.deleteFile(product.imageUrl)
         product.imageUrl = image.path
       }
       product.description = updatedDesc
@@ -179,7 +180,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId
-  Product.deleteOne({ _id: prodId, userId: req.user._id})
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'))
+      }
+      fileHelper.deleteFile(product.imageUrl)
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       res.redirect('/admin/products')
     })
